backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
supabase==2.10.0
python-dotenv==1.1.0
pydantic==2.11.3
